LifecycleMainActivity
Handled Lifecycle on main activity
