+++ app/src/main/java/com/example/popularmoviesapp/adapters/ReviewsAdapter.java
@@ -1,4 +1,4 @@
-package com.example.popularmoviesapp;
+package com.example.popularmoviesapp.adapters;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
@@ -7,6 +7,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.example.popularmoviesapp.R;
 import com.example.popularmoviesapp.models.Review;
 
 public class ReviewsAdapter extends RecyclerView.Adapter<ReviewsAdapter.ReviewsViewHolder> {
